Add support for --tun_dev option
Currently packetdrill (on Linux) hardcodes required tun device name to
be "tun0" and fails if it can't obtain that one.

This commit adds support for --tun_dev=<device name> config option,
which makes it possible to work around that requirement.

With special value "auto", whatever kernel picked for us will be used.
When using something else than "auto", it will try 10 times to obtain
the requested device name.
Not using this option at all is equivalent to --tun_name=tun0, which is
backward compatible behavior.

Signed-off-by: Jakub Jankowski <shasta@toxcorp.com>